Modified to write out all apertures, rather than just the chosen one. For backward compatibility and convenience, the chosen aperture still goes into the same named columns, and is repeated during the numbered sequence of other apertures.  Column names are the same but with the internal number for the aperture appended.
Removed now-redundant apflux and aprms columns, and split/renamed apoffsets
to offsets.

These changes break "update", it will need to be fixed to cope with the new
format.  **TODO**

It may also be desirable to add flags to disable various parts of this
behaviour.

Quirks:

The FLAG_NODP (bit 0) is set based on the chosen aperture.
I think this is sensible, but there may be strange situations where
the flag gets set while there are measurements available in other
apertures.

The OFF, RMS, EXTC and noise model in the RMS diagram all still use the
rcore aperture.