Changed buffer to store all apertures in one struct lc_point.  Many of the parameters in this structure were duplicates, and this was very wasteful of disk/memory.  Adjusted all processing to place aperture loops inside other loops, in order to minimise I/O when processing multiple apertures.
Added NROWMAST / nrows member of struct lc_mef to keep track of original
number of rows for detecting incompatible catalogues.

Adjusted read_ref to skip rows we know can't be used as comparison stars
when asked for a particular class in output (-c flag).  At the moment, this
is just class=-1 (stellar), due to the weird comparison star selection
in systematic.c, but may be able to be expanded if alternate comparison
star selection methods are used in the future.