Modified comparison star selection to use magnitude ranges on reference image, rather than individual images, when the -u flag is in use.  This behaviour is more sensible and corresponds better to what the user is likely to expect.  Effect on light curves seems to be minor.  This allows the set of stars we need to ensure always get written to the output file to cater for later uses of "update" to be reduced in size (sometimes significantly) - poor frames were tending to use quite different comparison stars due to the bias in their individual magnitude systems.
The linear transformations for astrometry are no longer computed or written
out when running with -c.  This is notionally because they won't be the same
(due to not having all the stars), but also to save time and space given
the purpose of the -c flag is really to make things run faster as much as
it is to reduce the size of the output files.